BUG: can't AssumeRole in a non ACTIVE account
When we run the import script, we list all the accounts regardless of
their state, and then try to DescribeCases in all of them. This is an
issue for non ACTIVE accounts because we won't be able to
sts:AssumeRole.
As a result, we must build a list of ACTIVE accounts only. This is what
this commit does.