MINOR: don't hardcode region in lambda arn
The arn of the lambda function for federated queries is currently
hardcoded to us-east-1. If one does not pay attention during pre
requisites the deployment of the centralization stack will fail, when
deployed in a different region. Let's make it more obvious that it needs
to be changed as well.